IOS-9339: Crash when login on a fresh installed app

+++ iMEGA/My Account/Launch/LaunchViewController.m
@@ -113,7 +113,7 @@
     if (request.type == MEGARequestTypeFetchNodes) {
         [self invalidateTimerAPI_EAGAIN];
         
-        float progress = request.transferredBytes / request.totalBytes;
+        float progress = (float)request.transferredBytes / (float)request.totalBytes;
         if (progress > 0 && progress <= 1.0) {
             //Avoid that the stroke goes back and forward.
             if (progress < self.circularShapeLayer.strokeEnd) {
